flsk/Reviews/stepik-get-to-the-balcony-main/requirements.txt
click==8.1.3
colorama==0.4.5
dominate==2.7.0
flake8==5.0.4
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.1
importlib-metadata==4.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
typing_extensions==4.3.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
